feat #50: 받은 후기의 별점 평균 계산하여 매너 지수에 적용

+++ src/main/java/com/gogoring/dongoorami/member/dto/response/MemberInfoResponse.java
@@ -17,11 +17,11 @@ public class MemberInfoResponse {
 
     private final String introduction;
 
-    private final Integer manner;
+    private final Double manner;
 
     @Builder
     public MemberInfoResponse(String nickname, String profileImage, String gender, Integer age,
-            String introduction, Integer manner) {
+            String introduction, Double manner) {
         this.nickname = nickname;
         this.profileImage = profileImage;
         this.gender = gender;
